+++ airflow_client/client/api/event_log_api.py
@@ -58,7 +58,10 @@ class EventLogApi(object):
         self.get_event_log_endpoint = _Endpoint(
             settings={
                 'response_type': (EventLog,),
-                'auth': [],
+                'auth': [
+                    'Basic',
+                    'Kerberos'
+                ],
                 'endpoint_path': '/eventLogs/{event_log_id}',
                 'operation_id': 'get_event_log',
                 'http_method': 'GET',
@@ -107,7 +110,10 @@ class EventLogApi(object):
         self.get_event_logs_endpoint = _Endpoint(
             settings={
                 'response_type': (EventLogCollection,),
-                'auth': [],
+                'auth': [
+                    'Basic',
+                    'Kerberos'
+                ],
                 'endpoint_path': '/eventLogs',
                 'operation_id': 'get_event_logs',
                 'http_method': 'GET',
