+++ airflow_client/client/api/role_api.py
@@ -58,7 +58,10 @@ class RoleApi(object):
         self.delete_role_endpoint = _Endpoint(
             settings={
                 'response_type': None,
-                'auth': [],
+                'auth': [
+                    'Basic',
+                    'Kerberos'
+                ],
                 'endpoint_path': '/roles/{role_name}',
                 'operation_id': 'delete_role',
                 'http_method': 'DELETE',
@@ -107,7 +110,10 @@ class RoleApi(object):
         self.get_role_endpoint = _Endpoint(
             settings={
                 'response_type': (Role,),
-                'auth': [],
+                'auth': [
+                    'Basic',
+                    'Kerberos'
+                ],
                 'endpoint_path': '/roles/{role_name}',
                 'operation_id': 'get_role',
                 'http_method': 'GET',
@@ -156,7 +162,10 @@ class RoleApi(object):
         self.get_roles_endpoint = _Endpoint(
             settings={
                 'response_type': (RoleCollection,),
-                'auth': [],
+                'auth': [
+                    'Basic',
+                    'Kerberos'
+                ],
                 'endpoint_path': '/roles',
                 'operation_id': 'get_roles',
                 'http_method': 'GET',
@@ -218,7 +227,10 @@ class RoleApi(object):
         self.patch_role_endpoint = _Endpoint(
             settings={
                 'response_type': (Role,),
-                'auth': [],
+                'auth': [
+                    'Basic',
+                    'Kerberos'
+                ],
                 'endpoint_path': '/roles/{role_name}',
                 'operation_id': 'patch_role',
                 'http_method': 'PATCH',
@@ -280,7 +292,10 @@ class RoleApi(object):
         self.post_role_endpoint = _Endpoint(
             settings={
                 'response_type': (Role,),
-                'auth': [],
+                'auth': [
+                    'Basic',
+                    'Kerberos'
+                ],
                 'endpoint_path': '/roles',
                 'operation_id': 'post_role',
                 'http_method': 'POST',
