+++ airflow_client/client/api/task_instance_api.py
@@ -43,7 +43,6 @@ from airflow_client.client.model.error import Error
 from airflow_client.client.model.extra_link_collection import ExtraLinkCollection
 from airflow_client.client.model.inline_response200 import InlineResponse200
 from airflow_client.client.model.list_task_instance_form import ListTaskInstanceForm
-from airflow_client.client.model.set_task_instance_note import SetTaskInstanceNote
 from airflow_client.client.model.task_instance import TaskInstance
 from airflow_client.client.model.task_instance_collection import TaskInstanceCollection
 from airflow_client.client.model.task_instance_reference import TaskInstanceReference
@@ -64,7 +63,10 @@ class TaskInstanceApi(object):
         self.get_extra_links_endpoint = _Endpoint(
             settings={
                 'response_type': (ExtraLinkCollection,),
-                'auth': [],
+                'auth': [
+                    'Basic',
+                    'Kerberos'
+                ],
                 'endpoint_path': '/dags/{dag_id}/dagRuns/{dag_run_id}/taskInstances/{task_id}/links',
                 'operation_id': 'get_extra_links',
                 'http_method': 'GET',
@@ -125,7 +127,10 @@ class TaskInstanceApi(object):
         self.get_log_endpoint = _Endpoint(
             settings={
                 'response_type': (InlineResponse200,),
-                'auth': [],
+                'auth': [
+                    'Basic',
+                    'Kerberos'
+                ],
                 'endpoint_path': '/dags/{dag_id}/dagRuns/{dag_run_id}/taskInstances/{task_id}/logs/{task_try_number}',
                 'operation_id': 'get_log',
                 'http_method': 'GET',
@@ -208,7 +213,10 @@ class TaskInstanceApi(object):
         self.get_mapped_task_instance_endpoint = _Endpoint(
             settings={
                 'response_type': (TaskInstance,),
-                'auth': [],
+                'auth': [
+                    'Basic',
+                    'Kerberos'
+                ],
                 'endpoint_path': '/dags/{dag_id}/dagRuns/{dag_run_id}/taskInstances/{task_id}/{map_index}',
                 'operation_id': 'get_mapped_task_instance',
                 'http_method': 'GET',
@@ -275,7 +283,10 @@ class TaskInstanceApi(object):
         self.get_mapped_task_instances_endpoint = _Endpoint(
             settings={
                 'response_type': (TaskInstanceCollection,),
-                'auth': [],
+                'auth': [
+                    'Basic',
+                    'Kerberos'
+                ],
                 'endpoint_path': '/dags/{dag_id}/dagRuns/{dag_run_id}/taskInstances/{task_id}/listMapped',
                 'operation_id': 'get_mapped_task_instances',
                 'http_method': 'GET',
@@ -414,7 +425,10 @@ class TaskInstanceApi(object):
         self.get_task_instance_endpoint = _Endpoint(
             settings={
                 'response_type': (TaskInstance,),
-                'auth': [],
+                'auth': [
+                    'Basic',
+                    'Kerberos'
+                ],
                 'endpoint_path': '/dags/{dag_id}/dagRuns/{dag_run_id}/taskInstances/{task_id}',
                 'operation_id': 'get_task_instance',
                 'http_method': 'GET',
@@ -475,7 +489,10 @@ class TaskInstanceApi(object):
         self.get_task_instances_endpoint = _Endpoint(
             settings={
                 'response_type': (TaskInstanceCollection,),
-                'auth': [],
+                'auth': [
+                    'Basic',
+                    'Kerberos'
+                ],
                 'endpoint_path': '/dags/{dag_id}/dagRuns/{dag_run_id}/taskInstances',
                 'operation_id': 'get_task_instances',
                 'http_method': 'GET',
@@ -603,7 +620,10 @@ class TaskInstanceApi(object):
         self.get_task_instances_batch_endpoint = _Endpoint(
             settings={
                 'response_type': (TaskInstanceCollection,),
-                'auth': [],
+                'auth': [
+                    'Basic',
+                    'Kerberos'
+                ],
                 'endpoint_path': '/dags/~/dagRuns/~/taskInstances/list',
                 'operation_id': 'get_task_instances_batch',
                 'http_method': 'POST',
@@ -653,7 +673,10 @@ class TaskInstanceApi(object):
         self.patch_mapped_task_instance_endpoint = _Endpoint(
             settings={
                 'response_type': (TaskInstanceReference,),
-                'auth': [],
+                'auth': [
+                    'Basic',
+                    'Kerberos'
+                ],
                 'endpoint_path': '/dags/{dag_id}/dagRuns/{dag_run_id}/taskInstances/{task_id}/{map_index}',
                 'operation_id': 'patch_mapped_task_instance',
                 'http_method': 'PATCH',
@@ -726,7 +749,10 @@ class TaskInstanceApi(object):
         self.patch_task_instance_endpoint = _Endpoint(
             settings={
                 'response_type': (TaskInstanceReference,),
-                'auth': [],
+                'auth': [
+                    'Basic',
+                    'Kerberos'
+                ],
                 'endpoint_path': '/dags/{dag_id}/dagRuns/{dag_run_id}/taskInstances/{task_id}',
                 'operation_id': 'patch_task_instance',
                 'http_method': 'PATCH',
@@ -777,148 +803,6 @@ class TaskInstanceApi(object):
                     'dag_run_id': 'path',
                     'task_id': 'path',
                     'update_task_instance': 'body',
-                },
-                'collection_format_map': {
-                }
-            },
-            headers_map={
-                'accept': [
-                    'application/json'
-                ],
-                'content_type': [
-                    'application/json'
-                ]
-            },
-            api_client=api_client
-        )
-        self.set_mapped_task_instance_note_endpoint = _Endpoint(
-            settings={
-                'response_type': (TaskInstance,),
-                'auth': [],
-                'endpoint_path': '/dags/{dag_id}/dagRuns/{dag_run_id}/taskInstances/{task_id}/{map_index}/setNote',
-                'operation_id': 'set_mapped_task_instance_note',
-                'http_method': 'PATCH',
-                'servers': None,
-            },
-            params_map={
-                'all': [
-                    'dag_id',
-                    'dag_run_id',
-                    'task_id',
-                    'map_index',
-                    'set_task_instance_note',
-                ],
-                'required': [
-                    'dag_id',
-                    'dag_run_id',
-                    'task_id',
-                    'map_index',
-                    'set_task_instance_note',
-                ],
-                'nullable': [
-                ],
-                'enum': [
-                ],
-                'validation': [
-                ]
-            },
-            root_map={
-                'validations': {
-                },
-                'allowed_values': {
-                },
-                'openapi_types': {
-                    'dag_id':
-                        (str,),
-                    'dag_run_id':
-                        (str,),
-                    'task_id':
-                        (str,),
-                    'map_index':
-                        (int,),
-                    'set_task_instance_note':
-                        (SetTaskInstanceNote,),
-                },
-                'attribute_map': {
-                    'dag_id': 'dag_id',
-                    'dag_run_id': 'dag_run_id',
-                    'task_id': 'task_id',
-                    'map_index': 'map_index',
-                },
-                'location_map': {
-                    'dag_id': 'path',
-                    'dag_run_id': 'path',
-                    'task_id': 'path',
-                    'map_index': 'path',
-                    'set_task_instance_note': 'body',
-                },
-                'collection_format_map': {
-                }
-            },
-            headers_map={
-                'accept': [
-                    'application/json'
-                ],
-                'content_type': [
-                    'application/json'
-                ]
-            },
-            api_client=api_client
-        )
-        self.set_task_instance_note_endpoint = _Endpoint(
-            settings={
-                'response_type': (TaskInstance,),
-                'auth': [],
-                'endpoint_path': '/dags/{dag_id}/dagRuns/{dag_run_id}/taskInstances/{task_id}/setNote',
-                'operation_id': 'set_task_instance_note',
-                'http_method': 'PATCH',
-                'servers': None,
-            },
-            params_map={
-                'all': [
-                    'dag_id',
-                    'dag_run_id',
-                    'task_id',
-                    'set_task_instance_note',
-                ],
-                'required': [
-                    'dag_id',
-                    'dag_run_id',
-                    'task_id',
-                    'set_task_instance_note',
-                ],
-                'nullable': [
-                ],
-                'enum': [
-                ],
-                'validation': [
-                ]
-            },
-            root_map={
-                'validations': {
-                },
-                'allowed_values': {
-                },
-                'openapi_types': {
-                    'dag_id':
-                        (str,),
-                    'dag_run_id':
-                        (str,),
-                    'task_id':
-                        (str,),
-                    'set_task_instance_note':
-                        (SetTaskInstanceNote,),
-                },
-                'attribute_map': {
-                    'dag_id': 'dag_id',
-                    'dag_run_id': 'dag_run_id',
-                    'task_id': 'task_id',
-                },
-                'location_map': {
-                    'dag_id': 'path',
-                    'dag_run_id': 'path',
-                    'task_id': 'path',
-                    'set_task_instance_note': 'body',
                 },
                 'collection_format_map': {
                 }
@@ -1633,164 +1517,4 @@ class TaskInstanceApi(object):
         kwargs['update_task_instance'] = \
             update_task_instance
         return self.patch_task_instance_endpoint.call_with_http_info(**kwargs)
-
-    def set_mapped_task_instance_note(
-        self,
-        dag_id,
-        dag_run_id,
-        task_id,
-        map_index,
-        set_task_instance_note,
-        **kwargs
-    ):
-        """Update the TaskInstance note.  # noqa: E501
-
-        Update the manual user note of a mapped Task Instance.  *New in version 2.5.0*   # noqa: E501
-        This method makes a synchronous HTTP request by default. To make an
-        asynchronous HTTP request, please pass async_req=True
-
-        >>> thread = api.set_mapped_task_instance_note(dag_id, dag_run_id, task_id, map_index, set_task_instance_note, async_req=True)
-        >>> result = thread.get()
-
-        Args:
-            dag_id (str): The DAG ID.
-            dag_run_id (str): The DAG run ID.
-            task_id (str): The task ID.
-            map_index (int): The map index.
-            set_task_instance_note (SetTaskInstanceNote): Parameters of set Task Instance note.
-
-        Keyword Args:
-            _return_http_data_only (bool): response data without head status
-                code and headers. Default is True.
-            _preload_content (bool): if False, the urllib3.HTTPResponse object
-                will be returned without reading/decoding response data.
-                Default is True.
-            _request_timeout (int/float/tuple): timeout setting for this request. If
-                one number provided, it will be total request timeout. It can also
-                be a pair (tuple) of (connection, read) timeouts.
-                Default is None.
-            _check_input_type (bool): specifies if type checking
-                should be done one the data sent to the server.
-                Default is True.
-            _check_return_type (bool): specifies if type checking
-                should be done one the data received from the server.
-                Default is True.
-            _host_index (int/None): specifies the index of the server
-                that we want to use.
-                Default is read from the configuration.
-            async_req (bool): execute request asynchronously
-
-        Returns:
-            TaskInstance
-                If the method is called asynchronously, returns the request
-                thread.
-        """
-        kwargs['async_req'] = kwargs.get(
-            'async_req', False
-        )
-        kwargs['_return_http_data_only'] = kwargs.get(
-            '_return_http_data_only', True
-        )
-        kwargs['_preload_content'] = kwargs.get(
-            '_preload_content', True
-        )
-        kwargs['_request_timeout'] = kwargs.get(
-            '_request_timeout', None
-        )
-        kwargs['_check_input_type'] = kwargs.get(
-            '_check_input_type', True
-        )
-        kwargs['_check_return_type'] = kwargs.get(
-            '_check_return_type', True
-        )
-        kwargs['_host_index'] = kwargs.get('_host_index')
-        kwargs['dag_id'] = \
-            dag_id
-        kwargs['dag_run_id'] = \
-            dag_run_id
-        kwargs['task_id'] = \
-            task_id
-        kwargs['map_index'] = \
-            map_index
-        kwargs['set_task_instance_note'] = \
-            set_task_instance_note
-        return self.set_mapped_task_instance_note_endpoint.call_with_http_info(**kwargs)
-
-    def set_task_instance_note(
-        self,
-        dag_id,
-        dag_run_id,
-        task_id,
-        set_task_instance_note,
-        **kwargs
-    ):
-        """Update the TaskInstance note.  # noqa: E501
-
-        Update the manual user note of a non-mapped Task Instance.  *New in version 2.5.0*   # noqa: E501
-        This method makes a synchronous HTTP request by default. To make an
-        asynchronous HTTP request, please pass async_req=True
-
-        >>> thread = api.set_task_instance_note(dag_id, dag_run_id, task_id, set_task_instance_note, async_req=True)
-        >>> result = thread.get()
-
-        Args:
-            dag_id (str): The DAG ID.
-            dag_run_id (str): The DAG run ID.
-            task_id (str): The task ID.
-            set_task_instance_note (SetTaskInstanceNote): Parameters of set Task Instance note.
-
-        Keyword Args:
-            _return_http_data_only (bool): response data without head status
-                code and headers. Default is True.
-            _preload_content (bool): if False, the urllib3.HTTPResponse object
-                will be returned without reading/decoding response data.
-                Default is True.
-            _request_timeout (int/float/tuple): timeout setting for this request. If
-                one number provided, it will be total request timeout. It can also
-                be a pair (tuple) of (connection, read) timeouts.
-                Default is None.
-            _check_input_type (bool): specifies if type checking
-                should be done one the data sent to the server.
-                Default is True.
-            _check_return_type (bool): specifies if type checking
-                should be done one the data received from the server.
-                Default is True.
-            _host_index (int/None): specifies the index of the server
-                that we want to use.
-                Default is read from the configuration.
-            async_req (bool): execute request asynchronously
-
-        Returns:
-            TaskInstance
-                If the method is called asynchronously, returns the request
-                thread.
-        """
-        kwargs['async_req'] = kwargs.get(
-            'async_req', False
-        )
-        kwargs['_return_http_data_only'] = kwargs.get(
-            '_return_http_data_only', True
-        )
-        kwargs['_preload_content'] = kwargs.get(
-            '_preload_content', True
-        )
-        kwargs['_request_timeout'] = kwargs.get(
-            '_request_timeout', None
-        )
-        kwargs['_check_input_type'] = kwargs.get(
-            '_check_input_type', True
-        )
-        kwargs['_check_return_type'] = kwargs.get(
-            '_check_return_type', True
-        )
-        kwargs['_host_index'] = kwargs.get('_host_index')
-        kwargs['dag_id'] = \
-            dag_id
-        kwargs['dag_run_id'] = \
-            dag_run_id
-        kwargs['task_id'] = \
-            task_id
-        kwargs['set_task_instance_note'] = \
-            set_task_instance_note
-        return self.set_task_instance_note_endpoint.call_with_http_info(**kwargs)
