+++ airflow_client/client/api_client.py
@@ -94,7 +94,7 @@ class ApiClient(object):
             self.default_headers[header_name] = header_value
         self.cookie = cookie
         # Set default User-Agent.
-        self.user_agent = 'OpenAPI-Generator/2.3.0/python'
+        self.user_agent = 'OpenAPI-Generator/2.5.1/python'
 
     def __enter__(self):
         return self
