+++ airflow_client/docs/DagWarningApi.md
@@ -33,6 +33,7 @@ List dag warnings
 
 ### Example
 
+* Basic Authentication (Basic):
 
 ```python
 import time
@@ -47,9 +48,19 @@ configuration = client.Configuration(
     host = "http://localhost/api/v1"
 )
 
+# The client must configure the authentication and authorization parameters
+# in accordance with the API server security policy.
+# Examples for each auth method are provided below, use the example that
+# satisfies your auth use case.
+
+# Configure HTTP basic authorization: Basic
+configuration = client.Configuration(
+    username = 'YOUR_USERNAME',
+    password = 'YOUR_PASSWORD'
+)
 
 # Enter a context with an instance of the API client
-with client.ApiClient() as api_client:
+with client.ApiClient(configuration) as api_client:
     # Create an instance of the API class
     api_instance = dag_warning_api.DagWarningApi(api_client)
     dag_id = "dag_id_example" # str | If set, only return DAG warnings with this dag_id. (optional)
@@ -85,7 +96,7 @@ Name | Type | Description  | Notes
 
 ### Authorization
 
-No authorization required
+[Basic](../README.md#Basic), [Kerberos](../README.md#Kerberos)
 
 ### HTTP request headers
